requirements.txt
-f https://download.pytorch.org/whl/cu113/torch_stable.html

tqdm==4.64.1
numpy==1.23.4
matplotlib==3.6.1
opencv-python==4.6.0.66
pyyaml==5.4.1

torch==1.10.2+cu113
torchvision==0.11.3+cu113
